Encrypted Memos (#3)
* A proposal to add encrypted memos to transaction outputs

* add minor comments

* more discussion around AEAD mode alternatives

* Update text/0002-encrypted-memos.md

Co-authored-by: Eran Rundstein <eran@rundste.in>

* Add discussion of memo size

* fix many formatting issues and minor issues

* rename file

* add references and improve formatting

* add an acknowledgment

* fix typo

* additionally commentary on memo size and the compressed commitment change

* update document to account for changes in release plan

after experimentation, we have determined that the best way to
release the change is to upgrade the servers first to support memos,
and then update clients later to actually write the memos.

because of this, we need to have an interim period where memos are
optional. ultimately, we want memos to be mandatory because it
creates output uniformity. the changes to the document make clear
that there is an interim period where the memos are optional and
later that they become mandatory, in order to completely document
the evolution of the protocol rules.

Co-authored-by: Eran Rundstein <eran@rundste.in>